+++ pywren_ibm_cloud/config.py
@@ -25,6 +25,7 @@ logger = logging.getLogger(__name__)
 COMPUTE_BACKEND_DEFAULT = 'ibm_cf'
 STORAGE_BACKEND_DEFAULT = 'ibm_cos'
 JOBS_PREFIX = "pywren.jobs"
+LOGS_PREFIX = "pywren.logs"
 RUNTIMES_PREFIX = "pywren.runtimes"
 EXECUTION_TIMEOUT = 595  # 595 seconds => 10 minutes
 MAX_AGG_DATA_SIZE = 4e6  # 4MB